Deleted samples DriveByEnoder and TeleopTank
Deleted 2 files:
iRadsTeleopTank_Iterative
iRadsAutoDriveByEncoder_linear

Because the current teamcode opmode iRadsAutoOpMode_Linear contains
the majority of the features of both sample opmodes, in additon to
Visual and Encoder Navigation code, continuing to have these
legacy sample opModes is more of a source of confusion than anything
else. However, the encoderDrive() method may be extracted later
and placed as a static method of a Utility class to serve as an
example.