EncoderNavigation: Provided absolute position calculation.
Transformation from Robot XY to game field XY.
getX(), getY(), getHeading(), getConfidence(),
getAbsoluteLocalizationAge()

-setPosition(X_mm,Y_mm,Heading_deg) accepts absolute position
on the field, and resets position confidence to 100.
-setSteps() without arguments will read encoder values directly.
It should be run as often as possible so it can accurately
integrate changes in encoder position over time.
-getAge() reports how many seconds since the last absolute
position update.
-getConfidence gives a number decending from the initial
confidence level. The diminishing confidence is calculated based
on getAge() and how much the robot drives. This confidence
calculation is currently arbitrary, but with some tuning it could
provide a threshold at which the robot seaks out an
absolute position reference.