debug: display beaconGoal coordinates. Fix goalHeading.
If interactiveInit.debugMode = ON, robot position 
goal(s) for beacon target will be displayed as 
goalX, goalY, and goalHeading. 1 or 2 coordinates will 
be displayed, depending on whether 
interactiveInit.beaconGoal == BOTH.

Names of beaconGoal(s) are displayed.

Corrected goalRobotHeading calculation. Now 
displays 90 for Blue team and 180 for Red team, 
which are the desired ROBOT headings for beacon 
activation.