Utility class given non-static members for getMotorTickRate()
Utility utilLeftLaunchSpeed initialized inside 
initialize() method. 
Utility Constructor provides 
ElapsedTime, DcMotor, and updatePeriod.
Once initialized, the motor speed in TicksPerSecond 
is return by the getMotorTickRate() method.

At present, the output is noisy, jumping around 
wildly.  This is probably because it the updatePeriod 
constant isn't used yet, so the update rate is around 
1000 Hz, so the denominator of the ticks/sec 
quotient is small and noisy.

Two bugs were fixed in the previous implementation:
getMotorTickRate() previously had an unused 
argument called 'elapsedTime,' which was instead 
replaced with 0 each time the method was called.
The second bug was that where this method was 
called, the current time: runtime.time() should have 
been used, but instead robot.periodSec() was used, 
which is just the length of the period.